A few more fixes for msvc-7.1
git-svn-id: http://svn.boost.org/svn/boost/trunk@63276 b8fc166d-592f-0410-95f2-cb63ce0dd405

+++ libs/random/test/statistic_tests.hpp
@@ -344,10 +344,12 @@ public:
   double probability(unsigned int r) const
   {
     if(r == classes()-1)
-      return 1-fac<double>(d)/std::pow(d, static_cast<double>(d+classes()-2))* 
+      return 1-fac<double>(d)/
+        std::pow(static_cast<double>(d), static_cast<double>(d+classes()-2)) *
         stirling2<double>(d+classes()-2, d);
     else
-      return fac<double>(d)/std::pow(d, static_cast<double>(d+r)) * 
+      return fac<double>(d)/
+        std::pow(static_cast<double>(d), static_cast<double>(d+r)) * 
         stirling2<double>(d+r-1, d-1);
   }
 private:
